Start reorganizing views
