+++ solver/src/main/java/org/chocosolver/solver/variables/IViewFactory.java
@@ -12,6 +12,11 @@ package org.chocosolver.solver.variables;
 import org.chocosolver.solver.ISelf;
 import org.chocosolver.solver.Model;
 import org.chocosolver.solver.variables.view.*;
+import org.chocosolver.solver.variables.view.bool.BoolNotView;
+import org.chocosolver.solver.variables.view.bool.BoolEqView;
+import org.chocosolver.solver.variables.view.bool.BoolLeqView;
+import org.chocosolver.solver.variables.view.integer.*;
+import org.chocosolver.solver.variables.view.set.*;
 
 import java.util.Arrays;
 
@@ -76,11 +81,11 @@ public interface IViewFactory extends ISelf<Model> {
             return ref().intVar(name, var.getValue() + cste);
         }
         if (ref().getSettings().enableViews()) {
-            int p = checkDeclaredView(var, cste, OffsetView.class);
+            int p = checkDeclaredView(var, cste, IntOffsetView.class);
             if(p>-1){
                 return var.getView(p).asIntVar();
             }else {
-                return new OffsetView(var, cste);
+                return new IntOffsetView(var, cste);
             }
         } else {
             int lb = var.getLB() + cste;
@@ -108,14 +113,14 @@ public interface IViewFactory extends ISelf<Model> {
             return ref().intVar(-var.getValue());
         }
         if (ref().getSettings().enableViews()) {
-            if (var instanceof MinusView) {
-                return ((MinusView) var).getVariable();
+            if (var instanceof IntMinusView) {
+                return ((IntMinusView) var).getVariable();
             } else {
-                int p = checkDeclaredView(var, -1, MinusView.class);
+                int p = checkDeclaredView(var, -1, IntMinusView.class);
                 if(p>-1){
                     return var.getView(p).asIntVar();
                 }else {
-                    return new MinusView(var);
+                    return new IntMinusView(var);
                 }
             }
         } else {
@@ -163,11 +168,11 @@ public interface IViewFactory extends ISelf<Model> {
             if (ref().getSettings().enableViews()) {
                 boolean rev = cste < 0;
                 cste = Math.abs(cste);
-                int p = checkDeclaredView(var, cste, ScaleView.class);
+                int p = checkDeclaredView(var, cste, IntScaleView.class);
                 if(p>-1){
                     return var.getView(p).asIntVar();
                 }else {
-                    v2 = new ScaleView(var, cste);
+                    v2 = new IntScaleView(var, cste);
                 }
                 if(rev){
                     v2 = intMinusView(v2);
@@ -259,11 +264,11 @@ public interface IViewFactory extends ISelf<Model> {
             return ref().boolVar(false);
         } else {
             if (ref().getSettings().enableViews()) {
-                int p = checkDeclaredView(x, c, EqView.class);
+                int p = checkDeclaredView(x, c, BoolEqView.class);
                 if (p >= 0) {
                     return x.getView(p).asBoolVar();
                 } else {
-                    return new EqView(x, c);
+                    return new BoolEqView(x, c);
                 }
             }else{
                 BoolVar b = ref().boolVar();
@@ -287,11 +292,11 @@ public interface IViewFactory extends ISelf<Model> {
             return ref().boolVar(true);
         } else {
             if (ref().getSettings().enableViews()) {
-                int p = checkDeclaredView(x, c, EqView.class);
+                int p = checkDeclaredView(x, c, BoolEqView.class);
                 if (p >= 0) {
                     return x.getView(p).asBoolVar().not();
                 } else {
-                    return new EqView(x, c).not();
+                    return new BoolEqView(x, c).not();
                 }
             } else {
                 BoolVar b = ref().boolVar();
@@ -315,11 +320,11 @@ public interface IViewFactory extends ISelf<Model> {
             return ref().boolVar(false);
         } else {
             if (ref().getSettings().enableViews()) {
-                int p = checkDeclaredView(x, c, LeqView.class);
+                int p = checkDeclaredView(x, c, BoolLeqView.class);
                 if (p >= 0) {
                     return x.getView(p).asBoolVar();
                 } else {
-                    return new LeqView(x, c);
+                    return new BoolLeqView(x, c);
                 }
             }else {
                 BoolVar b = ref().boolVar();
@@ -343,11 +348,11 @@ public interface IViewFactory extends ISelf<Model> {
             return ref().boolVar(false);
         } else {
             if (ref().getSettings().enableViews()) {
-                int p = checkDeclaredView(x, c - 1, LeqView.class);
+                int p = checkDeclaredView(x, c - 1, BoolLeqView.class);
                 if (p >= 0) {
                     return x.getView(p).asBoolVar().not();
                 } else {
-                    return new LeqView(x, c - 1).not();
+                    return new BoolLeqView(x, c - 1).not();
                 }
             }else {
                 BoolVar b = ref().boolVar();
@@ -360,25 +365,25 @@ public interface IViewFactory extends ISelf<Model> {
     static int checkDeclaredView(IntVar x, int c, Class clazz){
         for(int i = 0; i < x.getNbViews(); i++)
             if (clazz.isInstance(x.getView(i))) {
-                if(clazz  == EqView.class){
-                    EqView v = (EqView) x.getView(i);
+                if(clazz  == BoolEqView.class){
+                    BoolEqView v = (BoolEqView) x.getView(i);
                     if(v.cste == c){
                         return i;
                     }
-                }else if(clazz  == LeqView.class){
-                    LeqView v = (LeqView) x.getView(i);
+                }else if(clazz  == BoolLeqView.class){
+                    BoolLeqView v = (BoolLeqView) x.getView(i);
                     if(v.cste == c){
                         return i;
                     }
-                }else if(clazz == MinusView.class){
+                }else if(clazz == IntMinusView.class){
                     return i;
-                }else if(clazz == OffsetView.class){
-                    OffsetView v = (OffsetView) x.getView(i);
+                }else if(clazz == IntOffsetView.class){
+                    IntOffsetView v = (IntOffsetView) x.getView(i);
                     if(v.cste == c){
                         return i;
                     }
-                }else if(clazz == ScaleView.class){
-                    ScaleView v = (ScaleView) x.getView(i);
+                }else if(clazz == IntScaleView.class){
+                    IntScaleView v = (IntScaleView) x.getView(i);
                     if(v.cste == c){
                         return i;
                     }
@@ -467,7 +472,7 @@ public interface IViewFactory extends ISelf<Model> {
      * @return a set view such that boolVars[x - offset] = True <=> x in setView
      */
     default SetVar boolsSetView(BoolVar[] boolVars, int offset) {
-        return new BoolsSetView(offset, boolVars);
+        return new SetBoolsView(offset, boolVars);
     }
 
     // OVER ARRAY OF INTEGER VARIABLES
@@ -482,7 +487,7 @@ public interface IViewFactory extends ISelf<Model> {
      * @return a set view such that intVars[x - offset] = v[x - offset] <=> x in setView.
      */
     default SetVar intsSetView(IntVar[] intVars, int[] v, int offset) {
-        return new IntsSetView(v, offset, intVars);
+        return new SetIntsView(v, offset, intVars);
     }
 
     /**
@@ -528,7 +533,7 @@ public interface IViewFactory extends ISelf<Model> {
      * @return a set view over the set of nodes of a graph variable
      */
     default SetVar graphNodeSetView(GraphVar g) {
-        return new GraphNodeSetView(g);
+        return new SetNodeGraphView(g);
     }
 
     /**
@@ -538,7 +543,7 @@ public interface IViewFactory extends ISelf<Model> {
      * @return a set view over the set of successors of a node of a directed graph variable.
      */
     default SetVar graphSuccessorsSetView(DirectedGraphVar g, int node) {
-        return new GraphSuccessorsSetView(g, node);
+        return new SetSuccessorsGraphView(g, node);
     }
 
     /**
@@ -548,7 +553,7 @@ public interface IViewFactory extends ISelf<Model> {
      * @return a set view over the set of predecessors of a node of a directed graph variable.
      */
     default SetVar graphPredecessorsSetView(DirectedGraphVar g, int node) {
-        return new GraphPredecessorsSetView(g, node);
+        return new SetPredecessorsGraphView(g, node);
     }
 
     /**
@@ -558,6 +563,6 @@ public interface IViewFactory extends ISelf<Model> {
      * @return a set view over the set of neighbors of a node of an udirected graph variable.
      */
     default SetVar graphNeighborsSetView(UndirectedGraphVar g, int node) {
-        return new GraphSuccessorsSetView(g, node);
+        return new SetSuccessorsGraphView(g, node);
     }
 }
