+++ solver/src/main/java/org/chocosolver/solver/variables/view/bool/BoolLeqView.java
@@ -7,13 +7,14 @@
  *
  * See LICENSE file in the project root for full license information.
  */
-package org.chocosolver.solver.variables.view;
+package org.chocosolver.solver.variables.view.bool;
 
 import org.chocosolver.solver.ICause;
 import org.chocosolver.solver.exception.ContradictionException;
 import org.chocosolver.solver.learn.ExplanationForSignedClause;
 import org.chocosolver.solver.variables.IntVar;
 import org.chocosolver.solver.variables.events.IntEventType;
+import org.chocosolver.solver.variables.view.BoolIntView;
 import org.chocosolver.util.ESat;
 import org.chocosolver.util.objects.setDataStructures.iterable.IntIterableRangeSet;
 
@@ -26,7 +27,7 @@ import org.chocosolver.util.objects.setDataStructures.iterable.IntIterableRangeS
  * @author Charles Prud'homme
  * @since 04/02/11
  */
-public final class LeqView<I extends IntVar> extends IntBoolView {
+public final class BoolLeqView<I extends IntVar> extends BoolIntView {
 
     /**
      * A boolean view based on <i>var<i/> such that <i>var<i/> &le; <i>cste<i/>
@@ -34,7 +35,7 @@ public final class LeqView<I extends IntVar> extends IntBoolView {
      * @param var  an integer variable
      * @param cste an int
      */
-    public LeqView(final I var, final int cste) {
+    public BoolLeqView(final I var, final int cste) {
         super(var, "≤", cste);
     }
