+++ solver/src/main/java/org/chocosolver/solver/variables/view/integer/IntOffsetView.java
@@ -7,45 +7,49 @@
  *
  * See LICENSE file in the project root for full license information.
  */
-package org.chocosolver.solver.variables.view;
+package org.chocosolver.solver.variables.view.integer;
 
 import org.chocosolver.solver.ICause;
 import org.chocosolver.solver.exception.ContradictionException;
+import org.chocosolver.solver.learn.ExplanationForSignedClause;
 import org.chocosolver.solver.variables.IntVar;
 import org.chocosolver.solver.variables.delta.IIntDeltaMonitor;
 import org.chocosolver.solver.variables.delta.NoDelta;
-import org.chocosolver.solver.variables.events.IEventType;
 import org.chocosolver.solver.variables.events.IntEventType;
 import org.chocosolver.solver.variables.impl.scheduler.IntEvtScheduler;
+import org.chocosolver.solver.variables.view.IntView;
+import org.chocosolver.solver.variables.view.ViewDeltaMonitor;
 import org.chocosolver.util.iterators.DisposableRangeIterator;
 import org.chocosolver.util.iterators.DisposableValueIterator;
 import org.chocosolver.util.iterators.EvtScheduler;
+import org.chocosolver.util.objects.setDataStructures.iterable.IntIterableRangeSet;
 
-import static org.chocosolver.solver.variables.events.IntEventType.BOUND;
-import static org.chocosolver.solver.variables.events.IntEventType.DECUPP;
-import static org.chocosolver.solver.variables.events.IntEventType.INCLOW;
-import static org.chocosolver.solver.variables.events.IntEventType.INSTANTIATE;
 
 /**
- * View for -V, where V is a IntVar or view
- * <p>
- * <p>
+ * declare an IntVar based on X and C, such as X + C
+ * <br/>
  * Based on "Views and Iterators for Generic Constraint Implementations" <br/>
  * C. Shulte and G. Tack.<br/>
  * Eleventh International Conference on Principles and Practice of Constraint Programming
  *
  * @author Charles Prud'homme
- * @since 23/08/11
+ * @since 04/02/11
  */
-public class MinusView<I extends IntVar> extends IntView<I> {
-
+public final class IntOffsetView<I extends IntVar> extends IntView<I> {
 
     /**
-     * Create a -<i>var<i/> view
-     * @param var a integer variable
+     * A constant value
      */
-    public MinusView(final I var) {
-        super("-(" + var.getName() + ")", var);
+    public final int cste;
+
+    /**
+     * A view based on <i>var<i/> such that <i>var<i/> + <i>cste<i/>
+     * @param var an integer variable
+     * @param cste an int
+     */
+    public IntOffsetView(final I var, final int cste) {
+        super("(" + var.getName() + "+" + cste + ")", var);
+        this.cste = cste;
     }
 
     @Override
@@ -57,120 +61,87 @@ public class MinusView<I extends IntVar> extends IntView<I> {
         return new ViewDeltaMonitor(var.monitorDelta(propagator)) {
             @Override
             protected int transform(int value) {
-                return -value;
+                return value + cste;
             }
         };
     }
 
     @Override
-    public boolean updateBounds(int lb, int ub, ICause cause) throws ContradictionException {
-        assert cause != null;
-        int olb = this.getLB();
-        int oub = this.getUB();
-        boolean hasChanged = false;
-        if (olb < lb || oub > ub) {
-            IntEventType e = null;
-
-            if (olb < lb) {
-                model.getSolver().getEventObserver().updateLowerBound(this, lb, getLB(), cause);
-                e = INCLOW;
-                if (var.updateUpperBound(-lb, this)) {
-                    hasChanged = true;
-                } else {
-                    model.getSolver().getEventObserver().undo();
-                }
-            }
-            if (oub > ub) {
-                e = e == null ? DECUPP : BOUND;
-                model.getSolver().getEventObserver().updateUpperBound(this, ub, getUB(), cause);
-                if (var.updateLowerBound(-ub, this)) {
-                    hasChanged = true;
-                } else {
-                    model.getSolver().getEventObserver().undo();
-                }
-            }
-            if (isInstantiated()) {
-                e = INSTANTIATE;
-            }
-            if (hasChanged) {
-                this.notifyPropagators(e, cause);
-            }
-        }
-        return hasChanged;
-    }
-
-    @Override
     protected boolean doInstantiateVar(int value) throws ContradictionException {
-        return var.instantiateTo(-value, this);
+        return var.instantiateTo(value - cste, this);
     }
 
     @Override
     protected boolean doUpdateLowerBoundOfVar(int value) throws ContradictionException {
-        return var.updateUpperBound(-value, this);
+        return var.updateLowerBound(value - cste, this);
     }
 
     @Override
     protected boolean doUpdateUpperBoundOfVar(int value) throws ContradictionException {
-        return var.updateLowerBound(-value, this);
+        return var.updateUpperBound(value - cste, this);
     }
 
     @Override
     protected boolean doRemoveValueFromVar(int value) throws ContradictionException {
-        return var.removeValue(-value, this);
+        return var.removeValue(value - cste, this);
     }
 
     @Override
     protected boolean doRemoveIntervalFromVar(int from, int to) throws ContradictionException {
-        return var.removeInterval(-to, -from, this);
+        return var.removeInterval(from - cste, to - cste, this);
     }
 
     @Override
     public boolean contains(int value) {
-        return var.contains(-value);
+        return var.contains(value - cste);
     }
 
     @Override
     public boolean isInstantiatedTo(int value) {
-        return var.isInstantiatedTo(-value);
+        return var.isInstantiatedTo(value - cste);
     }
 
     @Override
     public int getValue() {
-        return -var.getValue();
+        return var.getValue() + cste;
     }
 
     @Override
     public int getLB() {
-        return -var.getUB();
+        return var.getLB() + cste;
     }
 
     @Override
     public int getUB() {
-        return -var.getLB();
+        return var.getUB() + cste;
     }
 
     @Override
     public int nextValue(int v) {
-        int value = var.previousValue(-v);
-        if (value == Integer.MIN_VALUE) return Integer.MAX_VALUE;
-        return -value;
+        int value = var.nextValue(v - cste);
+        if (value == Integer.MAX_VALUE) {
+            return value;
+        }
+        return value + cste;
     }
 
     @Override
     public int nextValueOut(int v) {
-        return -var.previousValueOut(-v);
+        return var.nextValueOut(v - cste) + cste;
     }
 
     @Override
     public int previousValue(int v) {
-        int value = var.nextValue(-v);
-        if (value == Integer.MAX_VALUE) return Integer.MIN_VALUE;
-        return -value;
+        int value = var.previousValue(v - cste);
+        if (value == Integer.MIN_VALUE) {
+            return Integer.MIN_VALUE;
+        }
+        return value + cste;
     }
 
     @Override
     public int previousValueOut(int v) {
-        return -var.nextValueOut(-v);
+        return var.previousValueOut(v - cste) + cste;
     }
 
     @Override
@@ -180,7 +151,7 @@ public class MinusView<I extends IntVar> extends IntView<I> {
 
     @Override
     public String toString() {
-        return "-(" + this.var.toString() + ") = [" + getLB() + "," + getUB() + "]";
+        return "(" + var.toString() + (cste >= 0 ? " + " : " - ") + Math.abs(cste) + ") = [" + getLB() + "," + getUB() + "]";
     }
 
     @Override
@@ -193,33 +164,33 @@ public class MinusView<I extends IntVar> extends IntView<I> {
                 @Override
                 public void bottomUpInit() {
                     super.bottomUpInit();
-                    vit = var.getValueIterator(false);
+                    vit = var.getValueIterator(true);
                 }
 
                 @Override
                 public void topDownInit() {
                     super.topDownInit();
-                    vit = var.getValueIterator(true);
+                    vit = var.getValueIterator(false);
                 }
 
                 @Override
                 public boolean hasNext() {
-                    return vit.hasPrevious();
-                }
-
-                @Override
-                public boolean hasPrevious() {
                     return vit.hasNext();
                 }
 
                 @Override
+                public boolean hasPrevious() {
+                    return vit.hasPrevious();
+                }
+
+                @Override
                 public int next() {
-                    return -vit.previous();
+                    return vit.next() + cste;
                 }
 
                 @Override
                 public int previous() {
-                    return -vit.next();
+                    return vit.previous() + cste;
                 }
 
                 @Override
@@ -247,43 +218,43 @@ public class MinusView<I extends IntVar> extends IntView<I> {
                 @Override
                 public void bottomUpInit() {
                     super.bottomUpInit();
-                    vir = var.getRangeIterator(false);
+                    vir = var.getRangeIterator(true);
                 }
 
                 @Override
                 public void topDownInit() {
                     super.topDownInit();
-                    vir = var.getRangeIterator(true);
+                    vir = var.getRangeIterator(false);
                 }
 
                 @Override
                 public boolean hasNext() {
-                    return vir.hasPrevious();
-                }
-
-                @Override
-                public boolean hasPrevious() {
                     return vir.hasNext();
                 }
 
                 @Override
-                public void next() {
-                    vir.previous();
+                public boolean hasPrevious() {
+                    return vir.hasPrevious();
                 }
 
                 @Override
-                public void previous() {
+                public void next() {
                     vir.next();
                 }
 
                 @Override
+                public void previous() {
+                    vir.previous();
+                }
+
+                @Override
                 public int min() {
-                    return -vir.max();
+                    return vir.min() + cste;
                 }
 
                 @Override
                 public int max() {
-                    return -vir.min();
+                    return vir.max() + cste;
                 }
 
                 @Override
@@ -302,30 +273,39 @@ public class MinusView<I extends IntVar> extends IntView<I> {
     }
 
     @Override
-    public IEventType transformEvent(IEventType evt) {
-        if (evt == INCLOW) {
-            return DECUPP;
-        } else if (evt == DECUPP) {
-            return INCLOW;
-        }
-        return evt;
-    }
-
-    @Override
     public void justifyEvent(IntEventType mask, int one, int two, int three) {
         switch (mask) {
             case DECUPP:
-                model.getSolver().getEventObserver().updateLowerBound(this, -one, -two, this);
+                model.getSolver().getEventObserver().updateUpperBound(this, one + cste, two + cste, this);
                 break;
             case INCLOW:
-                model.getSolver().getEventObserver().updateUpperBound(this, -one, -two, this);
+                model.getSolver().getEventObserver().updateLowerBound(this, one + cste, two + cste, this);
                 break;
             case REMOVE:
-                model.getSolver().getEventObserver().removeValue(this, -one, this);
+                model.getSolver().getEventObserver().removeValue(this, one + cste, this);
                 break;
             case INSTANTIATE:
-                model.getSolver().getEventObserver().instantiateTo(this, -one, this, -three, -two);
+                model.getSolver().getEventObserver().instantiateTo(this, one + cste, this, two + cste, three + cste);
                 break;
+        }
+    }
+
+    @Override
+    public void explain(int p, ExplanationForSignedClause explanation) {
+        IntIterableRangeSet set0, set1;
+        if (explanation.readVar(p) == this) { // case a. (see javadoc)
+            set1 = explanation.complement(getVariable());
+            set0 = explanation.domain(getVariable());
+            set0.plus(cste);
+            this.intersectLit(set0, explanation);
+            getVariable().unionLit(set1, explanation);
+        } else { // case b. (see javadoc)
+            assert explanation.readVar(p) == getVariable();
+            set0 = explanation.complement(this);
+            set1 = explanation.domain(this);
+            set1.minus(cste);
+            this.unionLit(set0, explanation);
+            getVariable().intersectLit(set1, explanation);
         }
     }
 }
