+++ solver/src/test/java/org/chocosolver/solver/variables/IntOffsetViewTest.java
@@ -27,7 +27,7 @@ import static org.testng.Assert.assertEquals;
  * @author Charles Prud'homme
  * @since 04/02/11
  */
-public class OffsetViewTest {
+public class IntOffsetViewTest {
 
     @Test(groups="1s", timeOut=60000)
     public void test1() {
